Preserve locations of some AST items and support outline selection update while the cursor is moving through the file.
Hopefully, under-cursor item localization is not affected (run tests!). I'm afraid of that
interleaving parent with children, like A + B, where "+" is a parent, may not find something
in case of binary search. Through, we seem to do a linear search.

It might be also  worth to compare LineCol as one 64-bit operation.